UnitTests -> IntegrationTests Extracting one builder source code generation to a separate class
